backend/requirements.txt
fastapi
uvicorn
pydantic[email][standard]
pydantic
pydantic-settings
motor
python-jose[cryptography]
passlib
bcrypt==4.0.1
pandas
numpy
scikit-learn
xgboost
tensorflow
shap
networkx
pytest
loguru
python-multipart
python-dotenv
